requirements-extra.txt
# FORMAT
# Put your extra requirements here in the following format
#
# package[version_required] # tag1, tag2, ...
#
# Allow at least one space between the package name/version and '#'
#
# Only one `#` is allowed per line. Lines starting with `#` are ignored.

coverage #                     dev
pytest==7.4   #                dev
flake8 #                       dev
sphinx #                       dev, doc
ipython #                      dev
numpydoc #                     dev, doc
sphinx-copybutton #            dev, doc
sphinx-panels #                dev, doc
pydata-sphinx-theme #          dev, doc
myst-parser>=0.15 #            dev, doc
myst-nb>=0.13 #                dev, doc
mdit-py-plugins>=0.3.1,<0.4 #  dev, doc
pre-commit #                   dev

pandapower #                   interop

# below work around a bug with tqdm bar with ipywidgets 8.0.1
ipywidgets==7.7 #              web

fastmcp>=2.0.0 #              mcp
pydantic>=2.0.0 #              mcp
requests>=2.31.0 #             mcp
beautifulsoup4>=4.12.0 #        mcp
